requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
tkinter

validators~=0.34.0
python-Wappalyzer~=0.3.1